feat: Add Dan Tri Xa Hoi RSS feed and documentation page
This commit adds the Dan Tri Xa Hoi RSS feed to the list of feeds to be fetched. It also includes the corresponding documentation page for the Xa Hoi category. This enhances the content of the website by providing news updates in the Xa Hoi category from the Dan Tri source.